Remove Unicode characters from XML files
With those characters present, the decompiler gives the following error message:

> Decompiler: Unable to initialize the DecompilerInterface: Could not register program:
> XML processing error: syntax error

Forcing YYERROR_VERBOSE and YYDEBUG to be 1 by editing Ghidra/Features/Decompiler/src/decompile/cpp/xml.cc makes the message barely more helpful:

> Decompiler: Unable to initialize the DecompilerInterface: Could not register program:
> XML processing error: syntax error, unexpected $end, expecting '-'

What's more annoying about this is that it wasn't the sla file, but instead one of the auxilary files (and it didn't specify it). I'm pretty sure it's the pspec file that was affected, as the ldefs one showed up in the UI properly, but I'm changing both just to be safe. It seems like Unicode in comments in the slaspec is still fine, though.