Restructure instructions
Similar to with GCDSP, this gives better results for instructions with multiple actions. In particular, it makes it possible to combine multiple opcodes separate from combining parameters, which is needed for parallel load/stores and conditionals. Unlike GCDSP, there is still a lot of code to create the root instruction table, but that's more feasible here since we have a few bits at the start of the instruction that determine its layout. Also, we can still have one operand per load/store, somewhat working around Ghidra's weird behavior with that (on GCDSP, the parameters for both the main and the extension opcode get treated as a single operand, which somewhat messes with xrefs).

This is a prerequisite to implementing looping.